+++ CHANGELOG.md
@@ -6,6 +6,11 @@ and this project try to adheres to [Semantic Versioning](https://semver.org/).
 Go to the `v2` branch to see the changelog of Lume 2.
 Go to the `v1` branch to see the changelog of Lume 1.
 
+## [3.1.5] - Unreleased
+### Changed
+- Simplified the logic to detect the hostname, port and location in `--serve` mode.
+- `LUME_PROXIED` environment variable no longer affects to the hostname, port and location detection.
+
 ## [3.1.4] - 2026-01-01
 ### Fixed
 - Removed a 10s await in the dev server [#803]
@@ -348,6 +353,7 @@ Go to the `v1` branch to see the changelog of Lume 1.
 [#797]: https://github.com/lumeland/lume/issues/797
 [#803]: https://github.com/lumeland/lume/issues/803
 
+[3.1.5]: https://github.com/lumeland/lume/compare/v3.1.4...HEAD
 [3.1.4]: https://github.com/lumeland/lume/compare/v3.1.3...v3.1.4
 [3.1.3]: https://github.com/lumeland/lume/compare/v3.1.2...v3.1.3
 [3.1.2]: https://github.com/lumeland/lume/compare/v3.1.1...v3.1.2